Altera extensão de arquivo
