--- src/App.js
@@ -1,30 +1,0 @@
-import React from "react";
-import { BrowserRouter, Routes, Route } from "react-router-dom";
-import TodoList from "./components/todoList/todoList";
-import Login from "./components/login/login";
-import Pomodoro from "./components/pomodoro/pomodoro";
-
-import { AuthProvider } from "./context/authContext";
-import { PomodoroProvider } from "./context/pomodoroContext";
-import NavBar from "./components/navBar/navBar";
-
-function App() {
-    return (
-        <div>
-            <BrowserRouter>
-                <AuthProvider>
-                    <PomodoroProvider>
-                        <NavBar />
-                        <Routes>
-                            <Route path="/" element={<TodoList />} />
-                            <Route path="/pomodoro" element={<Pomodoro />} />
-                            <Route path="/login" element={<Login />} />
-                        </Routes>
-                    </PomodoroProvider>
-                </AuthProvider>
-            </BrowserRouter>
-        </div>
-    );
-}
-
-export default App;
